5004: Rename of compaction-tracker v2 to v1
