+++ java/compaction/compaction-tracker/pom.xml
@@ -23,7 +23,7 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>compaction-trackerV2</artifactId>
+    <artifactId>compaction-tracker</artifactId>
 
     <dependencies>
         <dependency>
