+++ java/compaction/compaction-tracker/src/main/java/sleeper/compaction/tracker/job/CompactionJobTrackerFactory.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package sleeper.compaction.trackerv2.job;
+package sleeper.compaction.tracker.job;
 
 import software.amazon.awssdk.services.dynamodb.DynamoDbClient;
