+++ java/compaction/compaction-tracker/src/main/java/sleeper/compaction/tracker/task/DynamoDBCompactionTaskTrackerCreator.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.trackerv2.task;
+package sleeper.compaction.tracker.task;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
