+++ java/compaction/compaction-tracker/src/test/java/sleeper/compaction/trackerv2/task/StoreCompactionTaskIT.java
